Fix tests to reflect correct headcount numbers
When I completed this I was apparently naive to what "headcount" means.

What we want is not a count of the number of people hired in a given month, but
the number of people employed by the company at that date -- that is, a
cumulative sum of all hirings, less firings, up to that point.